9. Scattering, matte material, f32 -> f64

+++ src/interval.rs
@@ -1,18 +1,18 @@
 pub struct Interval {
-    pub min: f32,
-    pub max: f32
+    pub min: f64,
+    pub max: f64
 }
 
 impl Interval {
-    pub fn contains(self, x: f32) -> bool {
+    pub fn contains(self, x: f64) -> bool {
         self.min <= x && x <= self.max
     }
 
-    pub fn surrounds(self, x: f32) -> bool {
+    pub fn surrounds(self, x: f64) -> bool {
         self.min < x && x < self.max
     }
 
-    pub fn clamp(self, x: f32) -> f32 {
+    pub fn clamp(self, x: f64) -> f64 {
         if (x < self.min) {
             self.min
         } else if {x > self.max} {
@@ -34,5 +34,5 @@ impl std::clone::Clone for Interval { // interval.clone();
 
 impl std::marker::Copy for Interval {}
 
-pub const EMPTY: Interval = Interval{min: f32::INFINITY, max: -f32::INFINITY};
-pub const UNIVERSE: Interval = Interval{min: -f32::INFINITY, max: f32::INFINITY};
+pub const EMPTY: Interval = Interval{min: f64::INFINITY, max: -f64::INFINITY};
+pub const UNIVERSE: Interval = Interval{min: -f64::INFINITY, max: f64::INFINITY};
