+++ src/sphere.rs
@@ -5,22 +5,22 @@ use crate::interval::Interval;
 
 pub struct Sphere {
     pub center: Vec3,
-    pub radius: f32
+    pub radius: f64
 }
 
 impl Hittable for Sphere {
     fn hit(&self, r: Ray, t_range: Interval, rec: &mut HitRecord) -> bool {
         let oc: Vec3 = r.ori - self.center;
-        let a: f32 = r.dir.l2norm();
-        let half_b: f32 = oc*r.dir;
-        let c: f32 = oc.l2norm() - self.radius*self.radius;
-        let discriminant: f32 = half_b*half_b - a*c;
+        let a: f64 = r.dir.l2norm();
+        let half_b: f64 = oc*r.dir;
+        let c: f64 = oc.l2norm() - self.radius*self.radius;
+        let discriminant: f64 = half_b*half_b - a*c;
         if (discriminant < 0.) { 
             return false;
         }
 
-        let sqrt_d: f32 = discriminant.sqrt();
-        let root: f32 = (-half_b - sqrt_d)/a;                                   // Value of t for the nearest sphere hit 
+        let sqrt_d: f64 = discriminant.sqrt();
+        let root: f64 = (-half_b - sqrt_d)/a;                                   // Value of t for the nearest sphere hit 
         if (!t_range.surrounds(root)) {
             let root = (-half_b + sqrt_d)/a;
             if (!t_range.surrounds(root)) {
